requirements.txt
fastapi~=0.104.1
SQLAlchemy~=2.0.23
python-dotenv~=1.0.0